incois-platform-ui/backend/requirements.txt
fastapi[all]
psycopg2-binary
python-dotenv
transformers
torch
sentencepiece